feat(Cluster): add Tablets tab
