+++ src/containers/Cluster/Cluster.tsx
@@ -26,6 +26,7 @@ import {useTypedDispatch, useTypedSelector} from '../../utils/hooks';
 import {parseVersionsToVersionToColorMap} from '../../utils/versions';
 import {NodesWrapper} from '../Nodes/NodesWrapper';
 import {StorageWrapper} from '../Storage/StorageWrapper';
+import {TabletsTable} from '../Tablets/TabletsTable';
 import {Tenants} from '../Tenants/Tenants';
 import {Versions} from '../Versions/Versions';
 
@@ -152,6 +153,19 @@ export function Cluster({
                         error={clusterError}
                         additionalClusterProps={additionalClusterProps}
                     />
+                </Route>
+                <Route
+                    path={
+                        getLocationObjectFromHref(getClusterPath(clusterTabsIds.tablets)).pathname
+                    }
+                >
+                    <div className={b('tablets')}>
+                        <div className={b('fake-block')} />
+                        <TabletsTable
+                            tablets={cluster.SystemTablets ?? []}
+                            className={b('tablets-table')}
+                        />
+                    </div>
                 </Route>
                 <Route
                     path={
